Refactor: Organize API endpoints into modules
This commit reorganizes the API endpoints by moving them into dedicated modules within the backend/api/ directory. This improves code structure and maintainability.
