+++ backend/api/websocket.py
@@ -1,0 +1,57 @@
+"""WebSocket endpoints for real-time updates."""
+
+from fastapi import APIRouter, WebSocket, WebSocketDisconnect
+from typing import List
+
+router = APIRouter(tags=["websocket"])
+
+
+class ConnectionManager:
+    """WebSocket connection manager for real-time updates."""
+    def __init__(self):
+        self.active_connections: List[WebSocket] = []
+
+    async def connect(self, websocket: WebSocket):
+        """Accept a WebSocket connection."""
+        await websocket.accept()
+        self.active_connections.append(websocket)
+
+    def disconnect(self, websocket: WebSocket):
+        """Remove a WebSocket connection."""
+        if websocket in self.active_connections:
+            self.active_connections.remove(websocket)
+
+    async def send_personal_message(self, message: str, websocket: WebSocket):
+        """Send a message to a specific WebSocket connection."""
+        await websocket.send_text(message)
+
+    async def broadcast(self, message: str):
+        """Broadcast a message to all connected clients."""
+        for connection in self.active_connections:
+            try:
+                await connection.send_text(message)
+            except Exception:
+                # Remove disconnected connections
+                if connection in self.active_connections:
+                    self.active_connections.remove(connection)
+
+
+# Global WebSocket manager instance
+manager = ConnectionManager()
+
+
+@router.websocket("/ws")
+async def websocket_endpoint(websocket: WebSocket):
+    """
+    WebSocket endpoint for real-time updates.
+    
+    Connect to receive real-time notifications about events, suggestions,
+    and workflow executions.
+    """
+    await manager.connect(websocket)
+    try:
+        while True:
+            data = await websocket.receive_text()
+            await manager.send_personal_message(f"Message: {data}", websocket)
+    except WebSocketDisconnect:
+        manager.disconnect(websocket)
